+++ pages/creator-dashboard.js
@@ -3,6 +3,8 @@ import { useState, useEffect } from "react";
 import axios from "axios";
 import Web3Modal from "web3modal";
 
+import Layout from "../src/components/Layout";
+
 import { nftaddress, nftmarketaddress } from '../config';
 
 import NFT from '../artifacts/contracts/NFT.sol/NFT.json';
@@ -50,6 +52,7 @@ export default function CreatorDashboard() {
     }
     if (loadingState === 'loaded' && !nfts.length) return (<h1 className="py-10 px-20 text-3xl">No assets created</h1>)
     return (
+        <Layout>
         <div>
           <div className="p-4">
               <h2 className="text-2xl py-2">Items Created</h2>
@@ -90,5 +93,6 @@ export default function CreatorDashboard() {
               }
           </div>
       </div>
+      </Layout>
     )
 }
